requirements.txt
fastapi 
uvicorn
mysql-connector-python
pydantic